+++ test/Test.gd
@@ -8,6 +8,7 @@ static var ExampleStruct = NativePlugin.define_struct("ExampleStruct", {
 static var get_answer = NativePlugin.get_function("get_answer", "int")
 static var double_int = NativePlugin.get_function("double_int", "int", ["int"])
 static var double_float = NativePlugin.get_function("double_float", "float", ["float"])
+static var sum_ints = NativePlugin.get_function("sum_ints", "int", ["const int *", "int"])
 static var get_message = NativePlugin.get_function("get_message", "const char *")
 static var fill_message = NativePlugin.get_function("fill_message", "void", ["char *", "int"])
 static var get_example_struct = NativePlugin.get_function("get_example_struct", "ExampleStruct")
@@ -24,6 +25,10 @@ func _ready():
 	print(get_answer.invoke())
 	print(double_int.invoke(31))
 	print(double_float.invoke(2.51))
+	
+	var ints = PackedInt32Array([1, 2, 3, 4, 5])
+	assert(sum_ints.invoke(ints, ints.size()) == (1 + 2 + 3 + 4 + 5))
+	
 	var message: CFFIPointer = get_message.invoke()
 	var message_str = message.get_string_from_utf8()
 	print(message_str)